refactor: base class for record muxers
